tests/wpt/tests/html/canvas/offscreen/canvas-context/2d.canvas.context.extraargs.create.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.context.extraargs.create</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.context.extraargs.create</h1>
<p class="desc">The 2D context doesn't throw with extra getContext arguments (new context)</p>


<script>
var t = async_test("The 2D context doesn't throw with extra getContext arguments (new context)");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertDifferent(new OffscreenCanvas(100, 50).getContext('2d', false, {}, [], 1, "2"), null, "new OffscreenCanvas(100, 50).getContext('2d', false, {}, [], 1, \"2\")", "null");
  _assertDifferent(new OffscreenCanvas(100, 50).getContext('2d', 123), null, "new OffscreenCanvas(100, 50).getContext('2d', 123)", "null");
  _assertDifferent(new OffscreenCanvas(100, 50).getContext('2d', "test"), null, "new OffscreenCanvas(100, 50).getContext('2d', \"test\")", "null");
  _assertDifferent(new OffscreenCanvas(100, 50).getContext('2d', undefined), null, "new OffscreenCanvas(100, 50).getContext('2d', undefined)", "null");
  _assertDifferent(new OffscreenCanvas(100, 50).getContext('2d', null), null, "new OffscreenCanvas(100, 50).getContext('2d', null)", "null");
  _assertDifferent(new OffscreenCanvas(100, 50).getContext('2d', Symbol.hasInstance), null, "new OffscreenCanvas(100, 50).getContext('2d', Symbol.hasInstance)", "null");
  t.done();

});
</script>
